+++ CHANGELOG.md
@@ -4,7 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
-* The `indeterminate` mode for `flexberry-checkbox` by setting `isNullableBoolean`.
+* The `indeterminate` mode for `flexberry-checkbox` by setting `isNullable`.
 
 ## [2.6.0-beta.2] - 2020-12-20
 ### Added